Organize location components
